refactor: move LevelEvent to its own module
